fix inventaire ligne

+++ pharmaSmart-app/src/main/java/com/kobe/warehouse/service/inventaire/impl/InventaireQueryServiceImpl.java
@@ -21,15 +21,6 @@ import com.kobe.warehouse.service.inventaire.InventoryStockService;
 import com.kobe.warehouse.service.settings.AppConfigurationService;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.Tuple;
-import org.apache.commons.lang3.tuple.Triple;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedHashMap;
@@ -39,6 +30,14 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import org.apache.commons.lang3.tuple.Triple;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @Transactional(readOnly = true)
@@ -171,7 +170,9 @@ public class InventaireQueryServiceImpl implements InventaireQueryService {
         return new ArrayList<>(map.values());
     }
 
-    /** Stock théorique des seules lignes sans lot de l'export — une requête pour tout le lot */
+    /**
+     * Stock théorique des seules lignes sans lot de l'export — une requête pour tout le lot
+     */
     private Map<Integer, Integer> lotExportStockMap(Long inventoryId, List<Tuple> tuples) {
         Set<Integer> produitIds = tuples.stream()
             .filter(t -> t.get("inventory_lot_id", Long.class) == null)
@@ -237,7 +238,7 @@ public class InventaireQueryServiceImpl implements InventaireQueryService {
 
     private StoreInventoryLineRecord toRecord(Tuple t, Map<Integer, Integer> stockMap) {
         Integer produitId = t.get("produitId", Integer.class);
-        Long lotCount = t.get("lot_count", Long.class);
+        Integer lotCount = t.get("lot_count", Integer.class);
         return new StoreInventoryLineRecord(
             produitId,
             t.get("code_cip", String.class),
@@ -252,7 +253,7 @@ public class InventaireQueryServiceImpl implements InventaireQueryService {
             t.get("prix_uni", Integer.class),
             t.get("storage_id", Integer.class),
             t.get("seuil_mini", Integer.class),
-            lotCount != null ? lotCount.intValue() : 0,
+            lotCount != null ? lotCount : 0,
             t.get("classe_pareto", String.class),
             t.get("counted_by", String.class),
             t.get("updated_at", java.time.LocalDateTime.class),
@@ -278,26 +279,33 @@ public class InventaireQueryServiceImpl implements InventaireQueryService {
                 for (Tuple t : tuples) {
                     Number raw = (Number) t.get("famillyId");
                     long id = raw != null ? raw.longValue() : -1L;
-                    StoreInventoryLineExport line = StoreInventoryLineFilterBuilder.buildStoreInventoryLineExportRecord(t);
-                    map.put(id, buildFromTuple(map.get(id), id, line.getFamillyCode(), line.getFamillyLibelle(), line));
+                    StoreInventoryLineExport line = StoreInventoryLineFilterBuilder.buildStoreInventoryLineExportRecord(
+                        t);
+                    map.put(id, buildFromTuple(map.get(id), id, line.getFamillyCode(),
+                        line.getFamillyLibelle(), line));
                 }
             }
             case RAYON, NONE -> {
                 for (Tuple t : tuples) {
                     Number rawRayon = (Number) t.get("rayon_id");
                     Long idRayon = rawRayon != null ? rawRayon.longValue() : null;
-                    StoreInventoryLineExport line = StoreInventoryLineFilterBuilder.buildStoreInventoryLineExportRecord(t);
+                    StoreInventoryLineExport line = StoreInventoryLineFilterBuilder.buildStoreInventoryLineExportRecord(
+                        t);
                     Triple<Long, String, String> triple = getRayon(idRayon, line);
                     long id = triple.getLeft();
-                    map.put(id, buildFromTuple(map.get(id), id, triple.getMiddle(), triple.getRight(), line));
+                    map.put(id,
+                        buildFromTuple(map.get(id), id, triple.getMiddle(), triple.getRight(),
+                            line));
                 }
             }
             case STORAGE -> {
                 for (Tuple t : tuples) {
                     Number rawStorage = (Number) t.get("storage_id");
                     long id = rawStorage != null ? rawStorage.longValue() : -1L;
-                    StoreInventoryLineExport line = StoreInventoryLineFilterBuilder.buildStoreInventoryLineExportRecord(t);
-                    map.put(id, buildFromTuple(map.get(id), id, line.getStorageLibelle(), line.getStorageLibelle(), line));
+                    StoreInventoryLineExport line = StoreInventoryLineFilterBuilder.buildStoreInventoryLineExportRecord(
+                        t);
+                    map.put(id, buildFromTuple(map.get(id), id, line.getStorageLibelle(),
+                        line.getStorageLibelle(), line));
                 }
             }
         }
